requirements.txt
streamlit
python-dotenv
requests
openai
#pdf2image
pillow
pymupdf